+++ app/models/shipit/commit.rb
@@ -4,7 +4,7 @@ module Shipit
 
     belongs_to :stack, touch: true
     has_many :deploys
-    has_many :statuses, -> { order(created_at: :desc) }
+    has_many :statuses, -> { order(created_at: :desc) }, dependent: :destroy
     has_many :commit_deployments, dependent: :destroy
 
     after_commit { broadcast_update }